# examples/cactus-example-cbdc-bridging-backend/src/main/yml/openapi.yml
openapi: 3.0.3
info:
  title: CBDC-example backend API
  version: 0.0.2
  description: |
    Cactus-Example
  contact:
    name: Hyperledger Cacti
  x-logo:
    altText: Cacti logo
    backgroundColor: "#fafafa"
    url: https://github.com/hyperledger/cacti/blob/main/images/HL_Cacti_Logo_Color.png
...
paths:
  /api/v1/@hyperledger-cacti/cactus-example-cbdc/transact:
    post:
      tags:
        - transact
      x-hyperledger-cacti:
        http:
          verbLowerCase: post
          path: /api/v1/@hyperledger-cacti/cactus-example-cbdc/transact
      summary: Submit a transaction intent
      description: Allows users to queue intents for transactions based on specified parameters.
      operationId: Transact
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: ./schemas.yml#/TransactRequest
      responses:
        "200":
          description: Transaction successfully queued
          content:
            application/json:
              schema:
                $ref: ./schemas.yml#/TransactResponse
        default:
          description: An error occurred
          content:
            application/json:
              schema:
                $ref: ./schemas.yml#/APIError
...